Rename `helpers` group to `visibility`
