+++ frontend/test/__support__/e2e/commands/visibility/isVisibleInPopover.js
